LPS-112247 Rename ExportHelper to LayoutPageTemplatesExporter, and move to corresponding package
