+++ modules/apps/layout/layout-page-template-admin-web/src/main/java/com/liferay/layout/page/template/admin/web/internal/exporter/LayoutPageTemplatesExporter.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.layout.page.template.admin.web.internal.portlet.helper;
+package com.liferay.layout.page.template.admin.web.internal.exporter;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -65,8 +65,8 @@ import org.osgi.service.component.annotations.Reference;
 /**
  * @author Rubén Pulido
  */
-@Component(immediate = true, service = ExportHelper.class)
-public class ExportHelper {
+@Component(immediate = true, service = LayoutPageTemplatesExporter.class)
+public class LayoutPageTemplatesExporter {
 
 	public File exportDisplayPages(
 			List<LayoutPageTemplateEntry> layoutPageTemplateEntries)
@@ -347,7 +347,8 @@ public class ExportHelper {
 		}
 	}
 
-	private static final Log _log = LogFactoryUtil.getLog(ExportHelper.class);
+	private static final Log _log = LogFactoryUtil.getLog(
+		LayoutPageTemplatesExporter.class);
 
 	private static final ObjectMapper _objectMapper = new ObjectMapper() {
 		{
